Add basic session, login, logout

+++ mod1/sessions/solution/controllers/deleteUser.js
@@ -1,6 +1,10 @@
 const User = require("../models/user");
 
 module.exports = (req, res) => {
+  if (!req.session.userId) {
+    return res.send("Please login.");
+  }
+
   User.findOneAndDelete({ username: req.body.username }, (error, doc) => {
     if (error) {
       console.log(error);
